Feature/booking finish

+++ booking_service/src/services/address.py
@@ -5,7 +5,7 @@ from uuid import UUID
 
 from src.domain.dtos.address import AddressCreateDTO, AddressUpdateDTO
 from src.domain.entities.address import Address
-from src.services.exceptions import AddressNotFoundError
+from src.services.exceptions import AddressNotFoundError, ForbiddenError
 from src.services.interfaces.uow import IUnitOfWork
 
 logger = logging.getLogger(__name__)
@@ -16,16 +16,16 @@ class IAddressService(ABC):
     async def create_address(self, address: AddressCreateDTO) -> Address: ...
 
     @abstractmethod
-    async def get_address_by_id(self, address_id: UUID) -> Address: ...
+    async def get_address_by_id(self, address_id: UUID, user_id: UUID) -> Address: ...
 
     @abstractmethod
     async def get_my_addresses(self, user_id: UUID) -> Iterable[Address]: ...
 
     @abstractmethod
-    async def update_address(self, address: AddressUpdateDTO) -> Address: ...
+    async def update_address(self, address: AddressUpdateDTO, user_id: UUID, address_id: UUID) -> Address: ...
 
     @abstractmethod
-    async def delete_address(self, address_id: UUID) -> Address: ...
+    async def delete_address(self, address_id: UUID, user_id: UUID) -> Address: ...
 
 
 class AddressService(IAddressService):
@@ -43,7 +43,7 @@ class AddressService(IAddressService):
             created_address = await uow.address_repository.create(address)
             return created_address
 
-    async def get_address_by_id(self, address_id: UUID) -> Address:
+    async def get_address_by_id(self, address_id: UUID, user_id: UUID) -> Address:
         """
         Получает адрес по id.
         :param address_id: ID адреса.
@@ -55,6 +55,9 @@ class AddressService(IAddressService):
             if address is None:
                 logger.warning("Адрес с id=%s не найден", address_id)
                 raise AddressNotFoundError("Адрес не найден")
+            if address.user_id != user_id:
+                logger.warning("Адрес с id=%s не принадлежит пользователю с id=%s", address_id, user_id)
+                raise ForbiddenError("Адрес не принадлежит пользователю")
             return address
 
     async def get_my_addresses(self, user_id: UUID) -> Iterable[Address]:
@@ -68,7 +71,7 @@ class AddressService(IAddressService):
             addresses = await uow.address_repository.get_my_addresses(user_id=user_id)
             return addresses
 
-    async def update_address(self, address: AddressUpdateDTO) -> Address:
+    async def update_address(self, address: AddressUpdateDTO, user_id: UUID, address_id: UUID) -> Address:
         """
         Обновляет адрес.
         :param address: Адрес для обновления.
@@ -76,13 +79,20 @@ class AddressService(IAddressService):
         """
 
         async with self._uow as uow:
-            updated_address = await uow.address_repository.update(address=address)
+            current_address = await uow.address_repository.get_address(address_id=address_id)
+            if current_address is None:
+                logger.warning("Адрес с id=%s не найден", address_id)
+                raise AddressNotFoundError("Адрес не найден")
+            if current_address.user_id != user_id:
+                logger.warning("Адрес с id=%s не принадлежит пользователю с id=%s", address_id, user_id)
+                raise ForbiddenError("Адрес не принадлежит пользователю")
+            updated_address = await uow.address_repository.update(address=address, address_id=address_id)
             if updated_address is None:
-                logger.warning("Адрес с id=%s не найден", address.id)
+                logger.warning("Адрес с id=%s не найден", address_id)
                 raise AddressNotFoundError("Адрес не найден")
             return updated_address
 
-    async def delete_address(self, address_id: UUID) -> Address:
+    async def delete_address(self, address_id: UUID, user_id: UUID) -> Address:
         """
         Удаляет адрес.
         :param address_id: ID адреса.
@@ -90,6 +100,13 @@ class AddressService(IAddressService):
         """
 
         async with self._uow as uow:
+            current_address = await uow.address_repository.get_address(address_id=address_id)
+            if current_address is None:
+                logger.warning("Адрес с id=%s не найден", address_id)
+                raise AddressNotFoundError("Адрес не найден")
+            if current_address.user_id != user_id:
+                logger.warning("Адрес с id=%s не принадлежит пользователю с id=%s", address_id, user_id)
+                raise ForbiddenError("Адрес не принадлежит пользователю")
             deleted_address = await uow.address_repository.delete(address_id=address_id)
             if deleted_address is None:
                 logger.warning("Адрес с id=%s не найден", address_id)
